image_generation/requirements.txt
torch
torchvision
datasets
tqdm
Pillow
einops
numpy

